+++ software/tracksight/frontend/components/sdCard/SDCardDumpPage.tsx
@@ -1,0 +1,40 @@
+"use client";
+
+import { useEffect, useState } from "react";
+
+import useListSDCards from "@/lib/hooks/useListSDCards";
+
+import SDCardFileNavigator from "./SDCardFileNavigator";
+import SDCardPicker from "./SDCardPicker";
+
+const SDCardDumpPage = () => {
+  const [selectedSDCard, setSelectedSDCard] = useState<string | null>(null);
+  const { data: sdCards, isLoading, error } = useListSDCards();
+
+  useEffect(() => {
+    if (!sdCards || sdCards.length === 0) {
+      return;
+    }
+
+    if (!selectedSDCard || !sdCards.includes(selectedSDCard)) {
+      setSelectedSDCard(sdCards[0]);
+    }
+  }, [sdCards, selectedSDCard]);
+
+  return (
+    <>
+      <SDCardPicker
+        selected={selectedSDCard}
+        onSelect={setSelectedSDCard}
+        sdCards={sdCards}
+        isLoading={isLoading}
+        error={error}
+      />
+      <div className="flex-1 flex h-full" style={{ height: "100%" }}>
+        <SDCardFileNavigator sdCard={selectedSDCard} />
+      </div>
+    </>
+  );
+};
+
+export default SDCardDumpPage;
